Move `Vibration` interface to Switch
